Added a folder exists check in generate_test_data

+++ generate_test_data.py
@@ -7,8 +7,11 @@ episodes = 148
 
 def usecase1_test(show_name):
     
-    shutil.rmtree(show_name)
     p = Path(show_name)
+    
+    if p.exists():
+        shutil.rmtree(show_name)
+        
     p.mkdir()
     
     for i in range(1, episodes + 1):
